+++ pybdsim/Plot.py
@@ -514,8 +514,7 @@ def Histogram1DMultiple(histograms, labels, log=False, xlabel=None, ylabel=None,
     
     return f
 
-
-def Histogram2D(histogram, logNorm=False, xlogscale=False, ylocscale=False, xlabel="", ylabel="", zlabel="", title="", aspect="auto", scalingFactor=1.0, xScalingFactor=1.0, yScalingFactor=1.0, **imshowKwargs):
+def Histogram2D(histogram, logNorm=False, xLogScale=False, yLogScale=False, xlabel="", ylabel="", zlabel="", title="", aspect="auto", scalingFactor=1.0, xScalingFactor=1.0, yScalingFactor=1.0, **imshowKwargs):
     """
     Plot a pybdsim.Data.TH2 instance.
     logNorm        - logarithmic colour scale
@@ -541,9 +540,9 @@ def Histogram2D(histogram, logNorm=False, xlogscale=False, ylocscale=False, xlab
         _plt.imshow(sf*h.contents.T, extent=ext, origin='lower', aspect=aspect, **imshowKwargs)
         _plt.colorbar(format='%.0e', label=zlabel)
 
-    if xlogscale:
+    if xLogScale:
         _plt.xscale('log')
-    if ylocscale:
+    if yLogScale:
         _plt.yscale('log')
 
     if xlabel == "":